1. Fix databse to latest database name, database variable name and data type 2. Fix login issue for Buyer and Seller 3. Added in filter for Product Filtering (Product categories and Price Categories )

+++ src/main/java/com/example/ECommerce_BackEnd/model/additonalClasses/order_CustomerPastOrders.java
@@ -2,6 +2,7 @@ package com.example.ECommerce_BackEnd.model.additonalClasses;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -9,7 +10,7 @@ public class order_CustomerPastOrders {
     private int id_order;
     private Date order_date;
     private int product_id;
-    private int order_priceamount;
+    private BigDecimal order_priceamount;
     private String order_seller_username;
     private String order_description;
     private String order_status;
